Add code for 2.19
Phonebook:

Handle the case where we are trying to update a person that has already been
deleted from the server. Do this by readding the person we were trying to
modify.

Also fix another potential issue with simultaneous modification on two clients:
change the key for PersonTableItem components to person.id instead of
person.name, as names may be duplicated in the DB if there are simultaneous
additions. With this change, we use ids instead that are more likely to be
unique since the server takes care of keeping ids unique and all clients talk
to the same server. Non-unique keys cause errors and may lead to problems.