4.21: Bloglist backend: require token for delete
When deleting blogs, a valid token must now be passed in the Authorization
header. Otherwise a 401 status is returned.

At the same time, refactor token decoding a bit since it is now used in
multiple places. We don't want to repeat the same jwt.verify etc. process the
same way in multiple places - make it into a function instead.

Add a new test utility function createUserAndAuthenticate that can be used to
create a user and then immediately get a token for that user. This is used in
testing blog operations with authentication.

Modify tests for blog deletion to take into account the fact that a token is
now required. Add tests that ensure tokens are actually required and that
appropriate errors are shown when a user is unauthorized for whatever reason.

Add a new example request for blog deletion for debugging issues in development
related to blog deletion.