5.9: Bloglist frontend: add delete button
Make it possible to delete blogs from the list by adding a delete button. Also
extend the backend so that blogs with no associated user can be deleted by
anyone. Fix the frontend so that these kinds of blogs don't produce an error
but are instead shown as being added by an anonymous user.