6.3: Redux-anecdotes-v2: notification using store
Take the notification text of the Notification component from the Redux store.
To do this, add a new notificationReducer and make a combined reducer out of
the existing anecdoteReducer and the new reducer.

Also, fix the ESlint config so that no warning are displayed when running
"eslint .". Add a new "lint" script in package.json.