6.10&6.11: Redux-anecdotes-v2: backend+redux
Initialize anecdotes from a backend and add newly created anecdotes to the
backend. Create a new anecdote service that uses axios to communicate with a
backend. Use this service in App to initialize the list of anecdotes and in
AnecdoteForm to create new anecdotes. Modify the anecdote reducer so that it
has a new init action for anecdotes. Don't pass store to App anymore, as this
is no longer needed.